# [DO NOT MERGE] thread adjoint sparse matrix vector multiply