What is required for this to be mergeable?