@KristofferC 2018-10-05
And to provide a benchmark, using e.g. the Matrix in https://math.nist.gov/MatrixMarket/data/Harwell-Boeing/bcsstruc2/bcsstk16.html I get with 4 threads

```
julia> @btime mul!(c, M', x)
  75.789 μs (3 allocations: 96 bytes)
```

vs master

```
julia> @btime mul!(c, M', x)
  250.705 μs (2 allocations: 32 bytes)
```