@JeffBezanson 2018-10-05
Great test case for #22631.