@StefanKarpinski 2018-10-05
Am I understanding correctly that this

(a) already works on master and is (only) 50% slower than MKL?
(b) would potentially be faster (maybe competitive with MKL) after #22631 is merged?