@KristofferC 2018-10-05
a) yes
b) probably no, at least not in this case where there is no nested threading