@KristofferC 2018-10-05
> I guess the most important thing to parallellize is the standard mv-product (not the adjoint). 

Well, we should have used CSR then ;). But for algorithms dealing with symmetric matrices you could just call this one instead.