@KristofferC 2018-10-06
Because this is an isolated place with concrete julia Base types.  Broadcasting has user defined types and functions, seems 100x times more difficult and opens up a lot more space for different decisions.  

Anyway, I don't expect this to be merged right now but we have this cool threading framework and I just wanted to show a concrete place where it could be used for a good performance boost to get the discussion going.