@tknopp 2018-10-07
Broadcasting was just an example. Reductions could also benefit from multithreading. My concern was more that after #22631 the threading interface will certainly change.