@tknopp 2018-10-08
Yes thats true. Do you want this then as a hidden feature or should this be documented? (How to get a multithreaded sparse multiplication...)