@tknopp 2019-01-28
I have seen crashes when using nested threading. Is that what you mean by "interfere"? If yes, I am also for waiting for partr.