@KristofferC 2019-07-17
Update with results from master using

```jl
julia> using MatrixMarket

julia> M = MatrixMarket.mmread("bcsstk16.mtx");

julia> x = rand(size(M, 2)); y = similar(x);

julia> using BenchmarkTools

julia> using LinearAlgebra

julia> @btime mul!(y, M', x);
```

On master

```jl
julia> @btime mul!(y, M', x);
  250.861 μs (1 allocation: 16 bytes)
```

With this PR:

```jl

# 1 thread
julia> @btime mul!(y, M', x);
  281.209 μs (10 allocations: 816 bytes)

2 threads
julia> @btime mul!(y, M', x);
  153.064 μs (17 allocations: 1.48 KiB)

4 threads
julia> @btime mul!(y, M', x);
  88.616 μs (30 allocations: 2.86 KiB)
```

So there is a bit of overhead when using `@threads` and running with 1 thread.

Anything blocking something like this getting merged now with the new threading system in place?