@tknopp 2019-07-18
@KristofferC is nested threading now supported by the new system (without crashing but also without overcommitting threads)?