@KristofferC 2019-07-18
> In general, are we going to start threading other parts of stdlib or at least the sparse matrix implementation?

The point of the threading work surely gotta be that we want to start using it.

> The reason I ask is that it would otherwise be odd to have only one function multi-threaded - but then again, one has to start somewhere

Indeed you have to start somewhere. This seems like a simple place.

> @KristofferC is nested threading now supported by the new system (without crashing but also without overcommitting threads)?

Yes. That is the point of the partr scheduler work