@tknopp 2019-07-18
> Yes. That is the point of the partr scheduler work

I know, but was not sure of the threading macro already is adapted to not create too much tasks, if threading macros are nested.