@StefanKarpinski 2019-07-18
> but was not sure of the threading macro already is adapted to not create too much tasks, if threading macros are nested.

Tasks are cheap and are mapped onto hardware threads so it's ok to create lots of tasks and do so in a nested fashion—that's exactly how the system is supposed to work.