@jebej 2019-07-18
[From what I understood](https://discourse.julialang.org/t/julia-1-2-feature-freeze-april-4th/22478/15), `@threads` does not do nested threading yet.