@mbauman 2019-07-18
On current master, the `@threads` macro spins up as many tasks as there are threads and sticks each task to its own thread.  So it won't overcommit hardware threads, and it will happily work in a nested fashion (especially after #32477), but it also won't yet do the fancy depth-first scheduling that the non-sticky tasks do.