@jebej 2019-07-31
What's the performance currently?