@KristofferC 2019-07-31
The slowdown for 1 thread as said in https://github.com/JuliaLang/julia/pull/29525#issuecomment-512149371 might be worth discussing a bit more. Especially when we only run with 1 thread by default.