@KristofferC 2019-07-31
Also, I guess a `SparseVector` where the nonzeros are backed by a `BitArray` or something interesting like that is potentially dangerous.