@tkf 2019-08-25
It would be nice if threaded `mul!` is defined for transpose as well. I think one possibility is to do a refactoring like I proposed in #33069.