@KristofferC 2020-01-10
We introduced this nice new `AbstractSparseMatrixCSC`, so we can no longer assume that anything that uses it is threadsafe for anything (who knows what the implementation of such a matrix is made).