@ViralBShah 2020-01-10
It still has to be a CSC matrix, right? In which case, shouldn't be safe to multi-thread the operations. We are not updating the matrix itself, so I am not sure how thread safety affects this operation.