@KristofferC 2020-01-10
Thread safety is a property of the implementation and since it is `Abstract` who knows what anything does. I agree that it is unlikely to cause problems in practive though. However, https://github.com/JuliaLang/julia/pull/29525#issuecomment-517061447 might be worth thinking about.