@ViralBShah 2020-01-10
I still can't see how? I can imagine that is the output was an AbstractVector, it might matter. 

Is the worry that a different thread may be updating the AbstractSparseMatrix? If so, that worry would be relevant even in SparseMatrixCSC.

Agree that overhead of single threaded case with @threads is a real issue.