@tknopp 2020-01-10
In principle one could implement an `AbstractSparseMatrix` where the `colptr` (`nonzeros`/ `rowvals`) function returns an object which internally changes the representation when `getindex` is called. But that is pretty hypothetical. I don't see a use case for that.