@KristofferC 2020-01-10
You could do some caching on `getindex` calls so that the next `getindex` in the same column could be faster than a full binary search in some cases etc.