@tknopp 2020-01-10
I don't see a binary search there. The question is what the contract of `AbstractSparseMatrix` is. If the contract is `rowvals(A)::Vector` and `rowvals(A)::Vector` then its thread safe. If that contract does not hold (which is debatable) than it might not be thread-safe.