@KristofferC 2020-01-10
> I don't see a binary search there. 

My example wasn't very good. On `getindex` for a SparseMatrixCSC there is a binary search but for this code, there isn't.

> The question is what the contract of AbstractSparseMatrix is. If the contract is rowvals(A)::Vector and rowvals(A)::Vector then its thread safe. If that contract does not hold (which is debatable) than it might not be thread-safe.

Yeah, my point is mostly that this is something we are going to have to deal with on a larger scale when we want to start threading things more in Base. A lot of our code is quite generic and I don't think handwaving like "oh, it will probably be thread-safe for most reasonable implementations" is really going to cut it because people do and should be allowed to have "weird" implementations of e.g. `AbstractArray`.